% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANANSE_seurat_wrapper.R
\name{DEGS_scANANSE}
\alias{DEGS_scANANSE}
\title{DEGS_scANANSE}
\usage{
DEGS_scANANSE(
  seurat_object,
  min_cells = 50,
  output_dir = "~/",
  cluster_id = "seurat_clusters",
  additional_contrasts = "None"
)
}
\arguments{
\item{seurat}{object}

\item{min_cells:}{minimum of cells a cluster needs to be exported}

\item{output_dir:}{directory where the files are outputted}

\item{cluster_id:}{ID used for finding clusters of cells}

\item{additional_contrasts:}{additional contrasts to add between clusters within cluster_ID}
}
\description{
Calculate the differential genes needed for ananse influence
}
