% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANANSE_seurat_wrapper.R
\name{import_seurat_Maelstrom}
\alias{import_seurat_Maelstrom}
\title{import_seurat_Maelstrom}
\usage{
import_seurat_Maelstrom(
  seurat_object,
  cluster_id = "seurat_clusters",
  Maelstrom_dir = ""
)
}
\arguments{
\item{seurat_object}{object}

\item{cluster_id}{ID used for finding clusters of cells}

\item{Maelstrom_dir}{directory where the maelstrom results are stored}
}
\description{
load Maelstrom enriched motifs
}
