% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANANSE_seurat_wrapper.R
\name{import_seurat_scANANSE}
\alias{import_seurat_scANANSE}
\title{import_seurat_scANANSE}
\usage{
import_seurat_scANANSE(
  seurat_object,
  cluster_id = "seurat_clusters",
  anansnake_inf_dir = "None"
)
}
\arguments{
\item{seurat_object}{seurat object}

\item{cluster_id}{ID used for finding clusters of cells}

\item{anansnake_inf_dir}{influence directory generated by anansnake}
}
\description{
import the influences from a anansnake directory into a seurat object
}
