% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANANSE_seurat_wrapper.R
\name{Maelstrom_Motif2TF}
\alias{Maelstrom_Motif2TF}
\title{generate_TF_linked_motif_tables}
\usage{
Maelstrom_Motif2TF(
  seurat_object,
  mot_mat = NULL,
  m2f_df = NULL,
  cluster_id = "seurat_clusters",
  maelstrom_dir = "./maelstrom/",
  RNA_expression_assay = "RNA",
  RNA_expression_slot = "data",
  expr_tresh = 10,
  cor_tresh = 0.3,
  cor_method = "pearson",
  output_dir = "./",
  select_most_var = FALSE,
  return_df = F
)
}
\arguments{
\item{seurat_object}{object}

\item{mot_mat}{motif_matrix}

\item{cluster_id}{ID used for finding clusters of cells}

\item{maelstrom_dir}{directory where the GimmeMotifs m2f table is stored}

\item{RNA_expression_assay}{Seurat assay containing factor expression info}

\item{RNA_expression_slot}{slot within assay used for calculating mean factor expression per cluster}

\item{expr_tresh}{minimum sum of gene counts over all cells in RNA_expression_assay to filter genes by}

\item{cor_tresh}{minimum value of to filter the cor() output by}

\item{cor_method}{specify one of the cor() methods}

\item{output_dir}{where to store the results}

\item{return_df}{return both the seurat object and two dataframes with maelstrom scores and expression values as a list}

\item{all_motifs}{use all motifs (T), or only curated motifs (F)}

\item{combine_motifs}{mean (take mean multiple motifscores), max_var (take motif with highest variance), or cor (take motif with best correlation to gene expression)}
}
\description{
create motif-factor links & export tables for printing motif score alongside its binding factor
}
