% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANANSE_seurat_wrapper.R, R/export_functions.R
\name{export_ATAC_scANANSE}
\alias{export_ATAC_scANANSE}
\title{export_ATAC_scANANSE}
\usage{
export_ATAC_scANANSE(
  seurat_object,
  min_cells = 50,
  output_dir = "~/",
  ATAC_peak_assay = "peaks",
  cluster_id = "seurat_clusters"
)

export_ATAC_scANANSE(
  seurat_object,
  min_cells = 50,
  output_dir = "~/",
  ATAC_peak_assay = "peaks",
  cluster_id = "seurat_clusters"
)
}
\arguments{
\item{seurat_object}{object}

\item{min_cells}{minimum of cells a cluster needs to be exported}

\item{output_dir}{directory where the files are outputted}

\item{ATAC_peak_assay}{assay of the seurat object containing the peaks and peakcounts}

\item{cluster_id}{ID used for finding clusters of cells}
}
\description{
This functions exports ATAC values from a seurat object

This functions exports ATAC values from a seurat object
}
