% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANANSE_seurat_wrapper.R
\name{import_seurat_maelstrom}
\alias{import_seurat_maelstrom}
\title{import_seurat_Maelstrom}
\usage{
import_seurat_maelstrom(
  seurat_object,
  cluster_id = "seurat_clusters",
  maelstrom_dir = "~/",
  return_df = F
)
}
\arguments{
\item{seurat_object}{object}

\item{cluster_id}{ID used for finding clusters of cells}

\item{return_df}{return both the seurat object and a dataframe with maelstrom scores as a list}

\item{Maelstrom_dir}{directory where the maelstrom results are stored}
}
\description{
load Maelstrom enriched motifs
}
